+++ src/Core/Compiler.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2020 Ashar Khan <ashar786khan@gmail.com>
+ * Copyright (C) 2019-2021 Ashar Khan <ashar786khan@gmail.com>
  *
  * This file is part of CP Editor.
  *
@@ -47,7 +47,7 @@ class Compiler : public QObject
      * @brief destruct a compiler
      * @note the compilation process will be killed if it's running
      */
-    ~Compiler();
+    ~Compiler() override;
 
     /**
      * @brief start the compilation