+++ src/Core/Translator.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2020 Ashar Khan <ashar786khan@gmail.com>
+ * Copyright (C) 2019-2021 Ashar Khan <ashar786khan@gmail.com>
  *
  * This file is part of CP Editor.
  *
@@ -17,20 +17,23 @@
 
 #include "Core/Translator.hpp"
 #include "Core/EventLogger.hpp"
-#include "Settings/SettingsInfo.hpp"
+#include "generated/SettingsHelper.hpp"
 #include <QMap>
 #include <QTranslator>
 
 namespace Core
 {
-const static QMap<QString, QString> locales = {{"简体中文", "zh_CN"},
-                                               // {"繁體中文", "zh_TW"},
-                                               {"Русский", "ru_RU"}};
+const static QMap<QString, QString> locales = {{"简体中文", "zh_CN"}, {"Русский", "ru_RU"}};
+
+const static QMap<QString, QString> suffixes = {{"zh_CN", "_zh-CN"}, {"ru_RU", "_ru-RU"}};
+
+const static QMap<QString, QString> code = {{"zh_CN", "zh"}, {"ru_RU", "ru"}};
 
 QTranslator *Translator::translator = nullptr;
 
-void Translator::setLocale(const QString &language)
+void Translator::setLocale()
 {
+    const auto language = SettingsHelper::getLocale();
     LOG_INFO(INFO_OF(language));
     if (translator)
     {
@@ -57,5 +60,24 @@ void Translator::setLocale(const QString &language)
         LOG_ERR_IF(!qApp->installTranslator(translator), "Failed to load the translator " << translator);
         SettingsInfo::updateSettingInfo();
     }
+}
+
+QString Translator::langSuffix()
+{
+    return suffixes[langName()];
+}
+
+QString Translator::langCode()
+{
+    return code[langName()];
+}
+
+QString Translator::langName()
+{
+    const auto language = SettingsHelper::getLocale();
+
+    if (language == "system")
+        return QLocale::system().name();
+    return locales[language];
 }
 } // namespace Core