+++ src/Settings/PreferencesWindow.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2020 Ashar Khan <ashar786khan@gmail.com>
+ * Copyright (C) 2019-2021 Ashar Khan <ashar786khan@gmail.com>
  *
  * This file is part of CP Editor.
  *
@@ -47,7 +47,7 @@ class AddPageHelper
   public:
     explicit AddPageHelper(PreferencesWindow *w);
 
-    AddPageHelper &page(const QString &key, const QString &trkey, const QStringList &content);
+    AddPageHelper &page(const QString &key, const QString &trkey, const QStringList &content, bool alignTop = true);
     AddPageHelper &page(const QString &key, const QString &trkey, PreferencesPage *newpage);
     AddPageHelper &page(const QString &key, const QString &trkey, PreferencesPage *newpage, const QStringList &content);
 
@@ -59,6 +59,9 @@ class AddPageHelper
 
   private:
     bool atTop() const;
+
+    QString pathFor(const QString &key);
+    QString trPathFor(const QString &trkey);
 
     PreferencesWindow *window;
     QTreeWidget *tree;