+++ src/Telemetry/UpdateChecker.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2020 Ashar Khan <ashar786khan@gmail.com>
+ * Copyright (C) 2019-2021 Ashar Khan <ashar786khan@gmail.com>
  *
  * This file is part of CP Editor.
  *
@@ -49,7 +49,7 @@ class UpdateChecker : public QObject
     explicit UpdateChecker();
     void checkUpdate(bool silent);
     void closeAll();
-    ~UpdateChecker();
+    ~UpdateChecker() override;
 
   private slots:
     void cancelCheckUpdate();