+++ src/Util/Singleton.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2020 Ashar Khan <ashar786khan@gmail.com>
+ * Copyright (C) 2019-2021 Ashar Khan <ashar786khan@gmail.com>
  *
  * This file is part of CP Editor.
  *
@@ -35,13 +35,14 @@ template <typename T, typename D = T> class Singleton
         return inst;
     }
 
-  private:
-    Singleton() = default;
-    ~Singleton() = default;
     Singleton(const Singleton &) = delete;
     Singleton(Singleton &&) = delete;
     Singleton &operator=(const Singleton &) = delete;
     Singleton &operator=(Singleton &&) = delete;
+
+  private:
+    Singleton() = default;
+    ~Singleton() = default;
 };
 
 } // namespace Util