+++ src/Util/Util.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2020 Ashar Khan <ashar786khan@gmail.com>
+ * Copyright (C) 2019-2021 Ashar Khan <ashar786khan@gmail.com>
  *
  * This file is part of CP Editor.
  *
@@ -17,6 +17,10 @@
 
 #include "Util/Util.hpp"
 #include "Core/EventLogger.hpp"
+#include "Core/Translator.hpp"
+#include "generated/version.hpp"
+#include <QRegularExpression>
+#include <QUrl>
 #include <QWidget>
 
 namespace Util
@@ -28,6 +32,20 @@ void showWidgetOnTop(QWidget *widget)
     widget->setWindowState((widget->windowState() & ~Qt::WindowMinimized) | Qt::WindowActive);
     widget->activateWindow();
     widget->raise();
+}
+
+QString sanitizeAnchorName(const QString &str)
+{
+    return str.trimmed()
+        .remove(QRegularExpression(R"([!"\#$%&'()*+,\-./:;<=>?@\[\\\]^_‘{|}~]|，|（|）|。)"))
+        .toLower()
+        .replace(' ', '-');
+}
+
+QString websiteLink(const QString &path)
+{
+    return QUrl(QString("https://cpeditor.org/%1/%2/%3").arg(MINOR_VERSION).arg(Core::Translator::langCode()).arg(path))
+        .url(QUrl::NormalizePathSegments);
 }
 
 } // namespace Util