+++ src/mainwindow.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2020 Ashar Khan <ashar786khan@gmail.com>
+ * Copyright (C) 2019-2021 Ashar Khan <ashar786khan@gmail.com>
  *
  * This file is part of CP Editor.
  *
@@ -579,16 +579,7 @@ void MainWindow::applySettings(const QString &pagePath)
     if (pageChanged("Appearance/General"))
     {
         testcases->updateHeights();
-        if (SettingsHelper::isShowCompileAndRunOnly())
-        {
-            ui->compile->hide();
-            ui->runOnly->hide();
-        }
-        else
-        {
-            ui->compile->show();
-            ui->runOnly->show();
-        }
+        updateCompileAndRunButtons();
     }
 
     if (pageChanged("Appearance/Font"))
@@ -747,6 +738,7 @@ void MainWindow::setLanguage(const QString &lang)
     Util::applySettingsToEditor(editor, language);
     customCompileCommand.clear();
     ui->changeLanguageButton->setText(language);
+    updateCompileAndRunButtons();
     isLanguageSet = true;
     emit editorLanguageChanged(this);
 }
@@ -1296,6 +1288,31 @@ int MainWindow::timeLimit() const
     if (customTimeLimit == -1)
         return SettingsHelper::getDefaultTimeLimit();
     return customTimeLimit;
+}
+
+void MainWindow::updateCompileAndRunButtons() const
+{
+    if (language == "Python")
+    {
+        ui->runOnly->show();
+        ui->compile->hide();
+        ui->run->hide();
+    }
+    else
+    {
+        if (SettingsHelper::isShowCompileAndRunOnly())
+        {
+            ui->run->show();
+            ui->runOnly->hide();
+            ui->compile->hide();
+        }
+        else
+        {
+            ui->run->show();
+            ui->runOnly->show();
+            ui->compile->show();
+        }
+    }
 }
 
 // -------------------- COMPILER SLOTS ---------------------------